Update TestBBDSegment.java
The large new block is to handle the exception.  We wrap the 2 tests in a try/catch, and then we have another test to make sure that the exception is thrown when we expect it to be.

The block for new lines 261-263 are suggestions by the linter that made sense to me.  They are less verbose and do the same thing.

The hodge podge in the middle is due to some floating point drift in measuring the distance to points.  There were a few cases that ended up failing.  Not super sure why they weren't caught earlier.

Basically I am reinstantiating the vertical line every time, and allowing a delta for teh distance comparison.  Both due to that drift.