Make GameItem2d superficially compatible with BBDPolygon
A Polygon can be attached to the class on construction, and a flag can be set as to whetehr or not the Polygon needs to be updated during use.  Maybe the user just wants to programmatically create a shape and use that to render.  If so then they polygon will not need to be updated with translates etc.  if however the polygon needs to do things like measure dustance, area etc, then it needs to be accurate to the displayed form.