Convert double to float
Do this because JOML is requirwd unless I want to create a LOT of matrix math, and it is uses floats instead of doubles.

We lose a bit of range and precision.  We are now precies to within 1 ten thousandth of a unit, as opposed to the prior one in ten millionth.  This should still be sufficient.  The max is probably also a bit lower.  We are down from something silly like 10^300 to 10^38, which is still super large.