Create basic functions for calculating the intersection of 2 convex polygons
I expect the first 2 new ones to be used pretty extensively for all polygon operations.  The first one has uses beyond boolean operations, so it is public.  The other one really doesn't, so it will remain private.

These have NOT been tested yet.  That will be in the next commit.