Modify to return the game object
Do this because we want the object, even though its the mesh that we use. GameItems will need some sort of get collision mesh function that they can use for populating this, and to then allow the end users to make a simplified mesh to use for this sort of thing other than the fancy rendered one.